refactor: Create axios get requests for global data in App component to get the global data when app is initialized
