+++ src/App.tsx
@@ -7,8 +7,52 @@ import TeamPage from "./pages/TeamPage.tsx";
 import ContactPage from "./pages/ContactPage.tsx";
 import SignUpPage from "./pages/SignUpPage.tsx";
 import ProtectedPage from "./pages/ProtectedPage.tsx";
+import { AxiosError, AxiosResponse } from "axios";
+import { useAxios } from "./hooks/useAxios.tsx";
+import { useEffect } from "react";
+import { useAppDispatch } from "./store/store.ts";
+import { setGlobalData } from "./store/slices/globalSlice.ts";
 
 function App() {
+  const dispatch = useAppDispatch();
+  interface UserData {
+    roles: string[];
+    categories: string[];
+    language: string;
+    theme: string;
+  }
+  const [getData, getRequest]: [
+    AxiosResponse<UserData> | undefined,
+    (payload?: any, toastify?: boolean) => Promise<void>,
+    boolean,
+    AxiosError<any> | undefined
+  ] = useAxios({
+    reqType: "get",
+    endpoint: "roles",
+  });
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response: any = await getRequest();
+        const roles = response.map((role: any) => [role.name, role.code]);
+
+        dispatch(
+          setGlobalData({
+            roles: roles,
+            categories: [],
+            language: "tr",
+            theme: "lignt",
+          })
+        );
+        console.log(response);
+      } catch (error) {
+        console.error("Error fetching data:", error);
+      }
+    };
+    fetchData();
+  }, []);
+
   return (
     <>
       <Routes>
@@ -25,7 +69,7 @@ function App() {
           }
         />
         <Route
-          path="contact"
+          path="/contact"
           element={
             <ProtectedPage>
               <ContactPage />
