+++ src/components/SignUpForm.tsx
@@ -1,24 +1,11 @@
 import { Icon } from "@iconify/react/dist/iconify.js";
-
-import { useEffect, useState } from "react";
 import { useForm } from "react-hook-form";
 import { useLocation } from "react-router-dom";
 import { useAxios } from "../hooks/useAxios";
 import { AxiosError, AxiosResponse } from "axios";
+import { useAppSelector } from "../store/store";
+import { useState } from "react";
 
-type FormData = {
-  name: string;
-  email: string;
-  password: string;
-  confirmPassword: string;
-  phone: number;
-  birthDate: string;
-  role: string;
-  storeName: string;
-  storePhone: string;
-  storeTax: string;
-  storeBankAccount: string;
-};
 const SignUpForm = () => {
   const location = useLocation();
   const [postData, postRequest, postLoading, postError]: [
@@ -31,31 +18,11 @@ const SignUpForm = () => {
     endpoint: "signup",
     navPath: location.state ? location.state.pathname : "/",
   });
-  type RoleData = {
-    roles: string;
-  };
-  const [getData, getRequest]: [
-    AxiosResponse<RoleData> | undefined,
-    (payload?: any, toastify?: boolean) => Promise<void>,
-    boolean,
-    AxiosError<any> | undefined
-  ] = useAxios({
-    reqType: "get",
-    endpoint: "roles",
-  });
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        await getRequest();
-      } catch (error) {
-        console.error("Error fetching data:", error);
-      }
-    };
-    fetchData();
-  }, []);
+  const rolesArray = useAppSelector((state) => state.global.roles);
 
   const [hidePassword, setHidePassword] = useState<boolean>(true);
+
   function activateStoreDetails(e: React.ChangeEvent<HTMLSelectElement>) {
     const activeElement = document.getElementById("storeDetails");
     const role = e.target.value;
@@ -65,6 +32,20 @@ const SignUpForm = () => {
       activeElement?.classList.add("hidden");
     }
   }
+
+  type FormData = {
+    name: string;
+    email: string;
+    password: string;
+    confirmPassword: string;
+    phone: number;
+    birthDate: string;
+    role: string;
+    storeName: string;
+    storePhone: string;
+    storeTax: string;
+    storeBankAccount: string;
+  };
   const {
     register,
     handleSubmit,
@@ -92,13 +73,7 @@ const SignUpForm = () => {
 
     postRequest(submitData, true);
   };
-  useEffect(() => {
-    console.log("postdata", postData);
-  }, [postData]);
 
-  useEffect(() => {
-    console.log("get data", getData);
-  }, [getData]);
   return (
     <section
       id="signup-form"
@@ -277,9 +252,9 @@ const SignUpForm = () => {
               })}
               className={`w-full rounded-md border  bg-white py-3 px-6 mt-2 text-base font-medium text-[#6B7280] outline-none focus:border-sky-500 border-[#e0e0e0] focus:shadow-md `}
             >
-              {getData?.roles.map((role: any) => (
-                <option key={role} value={role.toLowerCase()}>
-                  {role}
+              {rolesArray.map((role: any) => (
+                <option key={role[1]} value={role[1]}>
+                  {role[0]}
                 </option>
               ))}
             </select>
